requirements.txt
openai==0.27.8
HTMLParser>=0.0.2
PyQRCode>=1.2.1
qrcode>=7.4.2
requests>=2.28.2
chardet>=5.1.0
Pillow
pre-commit
web.py

pymilvus~=2.3.4
python-dotenv~=1.0.0
websocket-client~=1.7.0
pydub~=0.25.1
pathlib~=1.0.1
pip~=23.3.2
transformers~=4.36.2